--- a/packages/lesswrong/client/start.tsx
+++ b/packages/lesswrong/client/start.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { hydrateRoot } from 'react-dom/client';
 import AppGenerator from './AppGenerator';
 import { onStartup } from '../lib/executionEnvironment';
 
@@ -21,12 +21,10 @@ onStartup(() => {
   const Main = () => (
     <AppGenerator apolloClient={apolloClient} abTestGroupsUsed={{}} />
   );
-
-  ReactDOM.hydrate(
-    <Main />,
-    document.getElementById('react-app'),
-    () => {
-      apolloClient.disableNetworkFetches = false;
-    }
-  );
+  
+  const container = document.getElementById('react-app');
+  const root = hydrateRoot(container!, <Main/>);
+  setTimeout(() => {
+    apolloClient.disableNetworkFetches = false;
+  }, 0);
 });