Upgrade from React 16 to 18. With this major version upgrade, they introduced concurrent rendering, which (hopefully) should greatly improve responsiveness on mobile during the window between when you see an SSR, and when it's fully hydrated.

Performance measurements from the Chrome profiler, client-side total scripting side, debug build, logged in front page, LW prod DB, Macbook Pro M1 Max:
* Before: 1575ms, 1140ms, 880ms, 819ms
* After: 831ms, 891ms, 684ms, 727ms

I have tested:
* Poking around the front page, post pages, etc works
* The post editor (CkEditor) still works
* Email rendering works (through `/debug/notificationEmailPreview`)

I also tried wrapping the tree in `<StrictMode>`; this gave a bunch of deprecation warnings, but none of the things the deprecation warnings pointed at seemed broken (yet).